Extract lens page renderers
