+++ dev-docs/module-reference.md
@@ -533,14 +533,23 @@ Route validation, per-profile last-view persistence, mobile tab sync handoff, an
 
 ---
 
+### `lens-pages.js`
+
+Dedicated page renderers for Labs, Genome, Body, Insight, and Recommendations. The module receives dashboard and recommendation helper functions from `views.js` through `createLensPageHandlers()` so page rendering can move out of the main views module without creating new import cycles.
+
+**Key exports:**
+- `createLensPageHandlers(deps)` — returns `showLabs`, `showGenomeLens`, `showBodyLens`, `showInsightLens`, and `showRecommendations`
+
+---
+
 ### `views.js`
 
-All dashboard, lens page, tool page, category, and modal rendering. The largest module. Public navigation remains exported here for compatibility, backed by `views-router.js`.
+Dashboard, Light page, tool page, category, and modal rendering. Public navigation and lens page functions remain exported here for compatibility, backed by `views-router.js` and delegated to `lens-pages.js` where applicable.
 
 **Key exports:**
 - `navigate(section, params)` — router facade created from `views-router.js`; calls the appropriate render function
 - `showDashboard(data?)` - renders the customizable widget dashboard; default widgets are Current Focus, Biological Age, Trends & Alerts, Recommended Next Steps, Marker Spotlight, Quick Markers, Biometrics Overview, Light Today, and Key Trends
-- `showLabs(data?)`, `showGenomeLens()`, `showBodyLens()`, `showInsightLens(data?)`, `showRecommendations(data?)` - dedicated lens/action pages
+- `showLabs(data?)`, `showGenomeLens()`, `showBodyLens()`, `showInsightLens(data?)`, `showRecommendations(data?)` - compatibility facades delegated to `lens-pages.js`
 - `showCompare(data?)` / `showCorrelations(data?)` - focused tool pages
 - Dashboard widget controls: `openDashboardWidgetPicker()`, `toggleDashboardOrganizeMode()`, `resetDashboardWidgets()`, `clearDashboardWidgets()`, `addDashboardWidgetFromLens()`, `removeDashboardWidgetFromLens()`
 - `showDetailModal(markerKey, data?)` — opens the marker detail modal
@@ -607,7 +616,7 @@ Client List modal for managing multiple profiles.
 
 ### `recommendations.js`
 
-Supplement, lifestyle, light-device, and EMF affiliate recommendations driven by a lazy-loaded catalog. Region-aware: products + URLs + coupons are filtered by user country via a single hierarchy chain (CZ → EU → INTL etc.). The global Recommendations page and dashboard widget live in `views.js`; this module owns catalog loading, slot rendering, disclosure state, product filtering, and DNA/wearable slot helpers.
+Supplement, lifestyle, light-device, and EMF affiliate recommendations driven by a lazy-loaded catalog. Region-aware: products + URLs + coupons are filtered by user country via a single hierarchy chain (CZ → EU → INTL etc.). The global Recommendations page is rendered by `lens-pages.js` and the dashboard widget helpers live in `views.js`; this module owns catalog loading, slot rendering, disclosure state, product filtering, and DNA/wearable slot helpers.
 
 **Key exports:**
 - `loadCatalog()` — lazy-loads `data/recommendations.json` on first call
